fix(context-offloader): keep optional @aws-sdk/client-s3 out of core Agent import graph
The core `Agent` statically imports `InMemoryStorage` from the context-offloader
`storage.ts`, which also housed `S3Storage`. `S3Storage` lazily loads the optional
`@aws-sdk/client-s3` peer via `await import(...)`, and a bundler eagerly resolves
string-literal dynamic imports while building the static graph. So importing only
`Agent` dragged the optional peer into the graph and broke esbuild builds for apps
that never use context offloading (the runtime `hasOffloader` guard fires too late).

Move `S3Storage` (the only consumer of `@aws-sdk/client-s3`) into its own
`s3-storage.ts`, re-exported unchanged from the `vended-plugins/context-offloader`
subpath. The core `Agent` import graph no longer reaches the S3 client, so the
optional peer is genuinely optional for bundlers. No public API change.

Adds a node-only regression test that bundles a `{ Agent }` entry with esbuild and
asserts `@aws-sdk/client-s3` is absent from the module graph (10 inputs -> 0).

Closes #3016
