+++ strands-ts/src/vended-plugins/context-offloader/index.ts
@@ -20,4 +20,5 @@
 export { ContextOffloader } from './plugin.js'
 export type { ContextOffloaderConfig } from './plugin.js'
 export type { Storage, FileStorageOptions } from './storage.js'
-export { InMemoryStorage, FileStorage, S3Storage } from './storage.js'
+export { InMemoryStorage, FileStorage } from './storage.js'
+export { S3Storage } from './s3-storage.js'
